# crates/pba_service/src/api/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "PurposeBoundAccountService",
        "version": "2026-04-14"
    },
    "paths": {
...
        "/accounts/{account_id}/deposits": {
            "post": {
                "operationId": "Deposit",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "deprecated": true,
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Deposit 201 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepositResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AccountNotFoundError 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountNotFoundErrorResponseContent"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "AccountNotActiveError 409 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountNotActiveErrorResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/accounts/{account_id}/withdrawals": {
            "post": {
                "operationId": "Withdraw",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WithdrawRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "deprecated": true,
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Withdraw 201 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WithdrawResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AccountNotFoundError 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountNotFoundErrorResponseContent"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "AccountNotActiveError 409 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountNotActiveErrorResponseContent"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "InsufficientFundsError 422 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InsufficientFundsErrorResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountNotActiveErrorResponseContent": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "error",
                    "message"
                ]
            },
            "AccountNotFoundErrorResponseContent": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "error",
                    "message"
                ]
            },
...
            "DepositRequestContent": {
                "type": "object",
                "properties": {
                    "source_ifsc": {
                        "type": "string"
                    },
                    "source_account_number": {
                        "type": "string"
                    },
                    "funding_type": {
                        "$ref": "#/components/schemas/FundingType"
                    },
                    "amount": {
                        "type": "number",
                        "description": "Monetary amount in the smallest currency unit (e.g., paise for INR)."
                    },
                    "pending": {
                        "type": "boolean"
                    },
                    "gateway_ref": {
                        "type": "string"
                    },
                    "timeout_seconds": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string",
                        "description": "Free-text narration for the deposit, stored on the transaction and\nreturned by the transaction listings. Max 256 characters."
                    },
                    "idempotency_key": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "source_account_number",
                    "source_ifsc"
                ]
            },
            "DepositResponseContent": {
                "type": "object",
                "properties": {
                    "deposit_id": {
                        "type": "string"
                    },
                    "account_id": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "description": "Monetary amount in the smallest currency unit (e.g., paise for INR)."
                    },
                    "pool": {
                        "type": "string"
                    },
                    "funding_type": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "gateway_ref": {
                        "type": "string"
                    },
                    "timeout_seconds": {
                        "type": "number"
                    }
                },
                "required": [
                    "account_id",
                    "amount",
                    "deposit_id",
                    "funding_type",
                    "pool",
                    "status"
                ]
            },
...
            "FundingType": {
                "type": "string",
                "description": "Funding source type for deposits.",
                "enum": [
                    "trust",
                    "third_party"
                ]
            },
...
            "InsufficientFundsErrorResponseContent": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "error",
                    "message"
                ]
            },
...
            "WithdrawRequestContent": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number",
                        "description": "Monetary amount in the smallest currency unit (e.g., paise for INR)."
                    },
                    "idempotency_key": {
                        "type": "string"
                    },
                    "gateway_ref": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "description": "Free-text narration for the withdrawal, stored on the transaction and\nreturned by the transaction listings. Max 256 characters."
                    }
                },
                "required": [
                    "amount"
                ]
            },
            "WithdrawResponseContent": {
                "type": "object",
                "properties": {
                    "account_id": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "description": "Monetary amount in the smallest currency unit (e.g., paise for INR)."
                    },
                    "gateway_ref": {
                        "type": "string"
                    }
                },
                "required": [
                    "account_id",
                    "amount"
                ]
            }
        },
        "securitySchemes": {
            "smithy.api.httpApiKeyAuth": {
                "type": "http",
                "description": "API key authentication via the 'Authorization' header",
                "name": "Authorization",
                "in": "header",
                "scheme": "ApiKey"
            }
        }
    },
    "security": [
        {
            "smithy.api.httpApiKeyAuth": []
        }
    ]
}